requirements.txt
streamlit
opencv-python
pandas
Pillow
rapidfuzz
